WatershedR: Prioritization of functional rare genetic variants by integrating genomic annotations and RNA-seq

Watershed is an unsupervised probabilistic framework that integrates genomic annotations and RNA-seq outlier calls to identify the probability a rare variant has a functional effect on a particular RNA-seq outlier phenotype (examples of outlier phenotypes can be, but are not limited to total expression, splicing, or ASE). Watershed extends our previous model RIVER (which can also be run via this package) by incoorperating information from multiple outlier phenoytpes into one model, where predictions for functional effects in one outlier phenotype are informed by observed outlier calls in another phenotype.

See also

Author

Maintainer: Nicole Gay nicole.r.gay@gmail.com [contributor]

Authors:

Ben Strober bstrober3@gmail.com